// src/interfaces/getLoggedUserData.ts
    export interface LoggedUserData {
            message: string;
            user: {
            _id: string;
            name: string;
            email: string;
            dateOfBirth: string; 
            gender: string;
            photo: string;
            createdAt: string; 
            passwordChangedAt: string; 
            };
        }